Move metrics provider interface to runtime package; add stubs for darwin
